refactor([issue-752]): extract appBuilder service from apps.js build route
Move the npm-install + build-command orchestration (and its
Windows shell-shim/kill-tree helpers) out of the 1180-LOC
server/routes/apps.js route handler into a dedicated
server/services/appBuilder.js. The route now validates the repo
path and delegates to appBuilder.buildApp(), which returns a
structured result the handler maps to HTTP status codes. Behavior,
error codes, and history logging are unchanged.
